Added a rate limiting feature to the TerminalProgressCallback.  This prevents it from updating too often, which will sometimes slow things down while the process blocks on console IO

+++ src/vw/Core/Thread.h
@@ -121,7 +121,7 @@ namespace vw {
         xt.sec++;
         milliseconds -= 1000;
       }
-      xt.nsec+=1e6*milliseconds;
+      xt.nsec+=int_fast32_t(1e6*milliseconds);
       boost::thread::sleep(xt);
     }
